test(gda): move the evidence verdicts onto the adapter and name the term
`tests/resource/test_import_evidence.py` asks the four evidence states of
`asset_state` and `project_import_gaps` directly, where each question costs a
temp tree and a call instead of the whole Typer + JSON-envelope toll: 18 verdict
tests move over unchanged in substance, and one new test pins
`classify_created_file` against the cache root for #839.

`test_resource_import_commands.py` keeps what the COMMAND decides — the pass
decision, the before/after accounting, the settlements, the refusals, the render
— plus one dry-run smoke per evidence state, so the wire ABI keeps its cover.
The artifact builders both modules need move to `tests/resource/import_artifacts`
so the two cannot drift on what a `cached` or `stale` tree looks like.

CONTEXT.md gains the `Import evidence` term.

Refs #741
